requirements.txt
# Core dependencies
adk==0.0.5
google-cloud-bigquery==3.11.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.0
google-cloud-firestore==2.11.0
google-cloud-documentai==2.20.0
google-cloud-functions==1.13.0
google-cloud-workflows==1.10.0

# Web framework
flask==2.3.3
fastapi==0.103.1
uvicorn==0.23.2

# Data processing
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1
